+++ R/default-options.R
@@ -1,8 +1,7 @@
 
 #' @importFrom scales hue_pal viridis_pal brewer_pal
-#' @importFrom hrbrthemes ipsum_pal ft_pal
 default_pals <- function() {
-  list(
+  pals <- list(
     choices = list(
       Default = list("ggplot2" = hue_pal()(9)),
       Viridis = list(
@@ -11,10 +10,6 @@ default_pals <- function() {
         "inferno" = viridis_pal(option = "inferno")(10),
         "plasma" = viridis_pal(option = "plasma")(10),
         "cividis" = viridis_pal(option = "cividis")(10)
-      ),
-      hrbrthemes = list(
-        "ipsum" = hrbrthemes::ipsum_pal()(9),
-        "ft" = hrbrthemes::ft_pal()(9)
       ),
       Diverging = list(
         "BrBG" = brewer_pal(palette = "BrBG")(11), 
@@ -59,52 +54,81 @@ default_pals <- function() {
       )
     ), 
     textColor = c(
-      rep(c("white", "black"), times = c(25, 18))
+      rep(c("white", "black"), times = c(23, 18))
     )
   )
+  
+  if (requireNamespace("hrbrthemes")) {
+    pals$choices$hrbrthemes = list(
+      "ipsum" = hrbrthemes::ipsum_pal()(9),
+      "ft" = hrbrthemes::ft_pal()(9)
+    )
+    pals$textColor<- c(
+      rep(c("white", "black"), times = c(25, 18))
+    )
+  }
+  return(pals)
 }
 
 
 #' @importFrom ggplot2 theme_bw theme_classic theme_dark theme_gray theme_grey 
 #'  theme_light theme_linedraw theme_minimal theme_void
-#' @importFrom ggthemes theme_base theme_calc theme_economist theme_economist_white 
-#'  theme_excel theme_few theme_fivethirtyeight theme_foundation theme_gdocs theme_hc 
-#'  theme_igray theme_map theme_pander theme_par theme_solarized theme_solarized_2 
-#'  theme_solid theme_stata theme_tufte theme_wsj
-#' @importFrom hrbrthemes theme_ft_rc theme_ipsum theme_ipsum_ps theme_ipsum_rc
-#'  theme_ipsum_tw theme_modern_rc
 default_themes <- function() {
-  list(
-    ggplot2 = list(
-      "bw", "classic", "dark", "gray",
-      "light", "linedraw", "minimal",
-      "void"
-    ),
-    hrbrthemes = c(
-      "ft_rc", "ipsum", "ipsum_ps", "ipsum_rc", "ipsum_tw", "modern_rc"
-    ),
-    ggthemes = list(
+  
+  ggplot2 <- c("bw", "classic", "dark", "gray",
+               "light", "linedraw", "minimal",
+               "void")
+  ggplot2 <- setNames(as.list(paste0("theme_", ggplot2)), ggplot2)
+  
+  themes <- list(
+    ggplot2 = ggplot2
+  )
+  if (requireNamespace("ggthemes")) {
+    ggthemes <- c(
       "base", "calc", "economist", "economist_white",
       "excel", "few", "fivethirtyeight", "foundation",
       "gdocs", "hc", "igray", "map", "pander",
       "par", "solarized", "solarized_2", "solid",
       "stata", "tufte", "wsj"
     )
-  )
+    ggthemes <- ggthemes[check_theme_exist(ggthemes, package = "ggthemes")]
+    if (length(ggthemes) > 0) {
+      ggthemes <- setNames(as.list(paste0("ggthemes::theme_", ggthemes)), ggthemes)
+      themes$ggthemes <- ggthemes
+    }
+  }
+  
+  if (requireNamespace("hrbrthemes")) {
+    hrbrthemes <- c(
+      "ft_rc", "ipsum", "ipsum_ps", "ipsum_rc", "ipsum_tw", "modern_rc"
+    )
+    hrbrthemes <- hrbrthemes[check_theme_exist(hrbrthemes, package = "hrbrthemes")]
+    if (length(hrbrthemes) > 0) {
+      hrbrthemes <- setNames(as.list(paste0("hrbrthemes::theme_", hrbrthemes)), hrbrthemes)
+      themes$hrbrthemes <- hrbrthemes
+    }
+  }
+  
+  return(themes)
+}
+
+check_theme_exist <- function(x, package = "hrbrthemes") {
+  vapply(X = paste0("theme_", x), FUN = function(fun) {
+    exists(fun, where = asNamespace(package), mode = "function")
+  }, FUN.VALUE = logical(1), USE.NAMES = FALSE)
 }
 
 
+
+#' @importFrom scales viridis_pal brewer_pal
 default_cols <- function() {
-  list(
+  cols <- list(
     "custom" = c("#0C4C8A", "#EF562D"),
     "viridis" = col2Hex(viridis_pal(option = "viridis")(10)),
     # "magma" = col2Hex(viridis_pal(option = "magma")(10)),
     # "inferno" = col2Hex(viridis_pal(option = "inferno")(10)),
     "plasma" = col2Hex(viridis_pal(option = "plasma")(10))#,
     # "cividis" = col2Hex(viridis_pal(option = "cividis")(10))
-    ,
-    "ipsum" = ipsum_pal()(9),
-    "ft" = ft_pal()(9)
     ,
     "Blues" = brewer_pal(palette = "Blues")(9),
     "Greens" = brewer_pal(palette = "Greens")(9),
@@ -116,4 +140,11 @@ default_cols <- function() {
     # "Set1" = brewer_pal(palette = "Set1")(8),
     # "Paired" = brewer_pal(palette = "Paired")(12)
   )
+  
+  if (requireNamespace("hrbrthemes")) {
+    cols$ipsum <- hrbrthemes::ipsum_pal()(9)
+    cols$ipsum <- hrbrthemes::ft_pal()(9)
+  }
+  
+  return(cols)
 }